% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m50_function.R
\name{m50}
\alias{m50}
\title{Multiply by 50}
\usage{
m50(x)
}
\arguments{
\item{x}{Value that you want to multiply}
}
\value{
x multiplied by 50
}
\description{
Multiply by 50
}
\examples{
m50(10)
[1] 500
}
