% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crw_function.R
\name{plot_crw}
\alias{plot_crw}
\title{Plot a correlated random walk}
\usage{
plot_crw(track, target_x = 30, target_y = 50)
}
\arguments{
\item{track}{Dataframe produced by simulate_crw()}

\item{target_x}{X coordinate of target point}

\item{target_y}{Y coordinate of target point}
}
\value{
A ggplot object
}
\description{
Plot a correlated random walk
}
