% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_crw_function.R
\name{simulate_crw}
\alias{simulate_crw}
\title{Simulate a correlated random walk}
\usage{
simulate_crw(
  n_steps = 100,
  step_length = 1,
  turn_sd = 0.3,
  target_x = 30,
  target_y = 50
)
}
\arguments{
\item{n_steps}{Number of movement steps}

\item{step_length}{Distance moved per step}

\item{turn_sd}{Standard deviation of turning angle}

\item{target_x}{X coordinate of target point}

\item{target_y}{Y coordinate of target point}
}
\value{
A dataframe containing movement coordinates
}
\description{
Simulate a correlated random walk
}
